Back-end/requirements.txt
astroid==2.4.2
blinker==1.4
cffi==1.14.0
click==7.1.2
cryptography==2.9.2
csvalidate==1.1.1
Flask==1.1.2
Flask-Cors==3.0.8
flask-marshmallow==0.13.0
Flask-SQLAlchemy==2.4.3
fpdf==1.7.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.6.1
marshmallow-sqlalchemy==0.23.1
mccabe==0.6.1
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
python-dotenv==0.13.0
python-http-client==3.2.7
sendgrid==6.3.1
six==1.15.0
SQLAlchemy==1.3.17
toml==0.10.1
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1
